maya: rename blocks/ → lists/, drop AnchorBlock + Toggle
The folder held only the KeyedList family once the legacy block helpers
fell away. Rename it to the honest category — lists/ — with an index.js
barrel (pure barrel: co-equal leaves, no namesake primary). lists/ also
gives VirtualList a home when it lands.

- delete AnchorBlock.js (orphan) and Toggle.js (only the barrel
  re-exported it; no importers)
- subpath export ./blocks -> ./lists; azoth umbrella re-export follows
- valhalla import + the live KeyedList design-doc link updated
- TODO: drop the resolved AnchorBlock/Toggle line

Full suite green: 301 passed | 4 skipped.
