+++ packages/azoth/maya.js
@@ -1,3 +1,3 @@
 export * from '@azothjs/maya/compose';
+export * from '@azothjs/maya/lists';
 export * from '@azothjs/maya/renderer';
-export * from '@azothjs/maya/blocks';
