--- packages/maya/blocks/Toggle.js
@@ -1,3 +1,0 @@
-export function Toggle({ on: predicate }, childNodes) {
-    return payload => predicate(payload) ? childNodes : null;
-}
