+++ packages/maya/lists/index.js
@@ -1,4 +1,3 @@
-export * from './Toggle.js';
 export * from './KeyedList.js';
 export * from './KeyedUList.js';
 export * from './KeyedOList.js';
