+++ packages/maya/package.json
@@ -26,9 +26,9 @@
         "**/!(*test|*config).js"
     ],
     "exports": {
-        "./blocks": "./blocks/blocks.js",
         "./channels": "./channels/index.js",
         "./compose": "./compose/compose.js",
+        "./lists": "./lists/index.js",
         "./renderer": "./renderer/renderer.js"
     },
     "devDependencies": {
